+++ universal-login-provider/src/ULIFrameProvider.ts
@@ -1,5 +1,6 @@
 import {RpcBridge} from './RpcBridge';
 import {DEFAULT_CONFIG} from './config'
+import {createIFrame} from './createIframe';
 
 export class ULIFrameProvider {
   private iframe: HTMLIFrameElement;
@@ -35,18 +36,3 @@ export class ULIFrameProvider {
   }
 }
 
-function createIFrame(url: string) {
-  const iframe = document.createElement('iframe');
-  Object.assign(iframe.style, {
-    // position: 'fixed',
-    width: '100vw',
-    height: '100vh',
-    left: '0',
-    top: '0',
-    background: 'none transparent',
-    border: 'none',
-  })
-  iframe.setAttribute('src', url);
-  document.getElementsByTagName('body')[0].appendChild(iframe);
-  return iframe;
-}
