+++ universal-login-provider/src/createIframe.ts
@@ -1,0 +1,15 @@
+export function createIFrame(url: string) {
+  const iframe = document.createElement('iframe');
+  Object.assign(iframe.style, {
+    width: '100vw',
+    height: '100vh',
+    left: '0',
+    top: '0',
+    background: 'none transparent',
+    border: 'none',
+  })
+  iframe.setAttribute('src', url);
+  document.getElementsByTagName('body')[0].appendChild(iframe);
+  return iframe;
+}
+
